+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+## [0.0.2]
+### Added
 - Support for InChIKeys
+- Support for exact and average masses
 
 ### Changed
 - Easier local development (just rebuild and restart container)
@@ -12,6 +25,7 @@
 ### Removed
 
 ### Fixed
+- Cache is now putting values in REDIS
 
 ### Security
 ## [0.0.1]